refactoring to support for custom properties provider selector
